Update tileMapEditorUtilities.js
Fixed bug where if user clicks on New map, then hovers mouse over a tile in Search then hovers away, the tile preview hover on the map does not clear. This is because
"let layeredTileHashesDisplay = tileMapEditorData.layeredTileHashesDisplay;" is set above the function. New map changes the reference of layeredTileHashesDisplay in tileMapEditorData.